+++ src/main/kotlin/com/example/hhplus_ecommerce/domain/payment/event/PaymentEventListener.kt
@@ -1,8 +1,7 @@
-package com.example.hhplus_ecommerce.domain.payment
+package com.example.hhplus_ecommerce.domain.payment.event
 
-import com.example.hhplus_ecommerce.event.DataPlatformEvent
-import com.example.hhplus_ecommerce.infrastructure.message.producer.MessageProducer
-import com.example.hhplus_ecommerce.infrastructure.message.producer.dto.PaymentDataMessage
+import com.example.hhplus_ecommerce.infrastructure.kafka.producer.MessageProducer
+import com.example.hhplus_ecommerce.infrastructure.kafka.producer.dto.PaymentDataMessage
 import org.springframework.stereotype.Component
 import org.springframework.transaction.event.TransactionPhase
 import org.springframework.transaction.event.TransactionalEventListener
@@ -11,8 +10,9 @@ import org.springframework.transaction.event.TransactionalEventListener
 class PaymentEventListener(
     private val messageProducer: MessageProducer
 ) {
+
     @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-    fun listen(event: DataPlatformEvent) {
+    fun listen(event: AfterPaymentEvent) {
         // 외부 데이터 플랫폼에 전송하는 이벤트 발행
         messageProducer.sendPaymentDataPlatformMessage(PaymentDataMessage.from(event.paymentResultInfo))
     }
